Move root package to com.example since otherwise cloudant configuration properties bean is multiple defined
